requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
groq>=0.4.0
python-dotenv>=1.0.1
rich>=13.7.1
